// src/services/slices/constructorSlice/types.ts
import { TConstructorIngredient, TOrder } from '@utils-types';

export interface constructorState {
  isLoading: boolean;
  constructorItems: {
    bun: TConstructorIngredient | null;
    ingredients: TConstructorIngredient[];
  };
  orderRequest: boolean;
  orderModalData: TOrder | null;
  error: string | null;
}